Correction d'une désinchronisation des objects dû à la séparations des thread via newt. Les angles de rotation sont maintenant calculés en amont et plus sommé via les transformations matricielles.

+++ src/fr/legbt/sections/Volume.java
@@ -57,19 +57,19 @@ public abstract class Volume {
 		}
 	}
 
-	public void xRotation(float degree){
+	public void xRotation(double degree){
 		for(int i=0;i<pieces.size();i++){
-			pieces.get(i).xRotation(degree);
+			pieces.get(i).xRotation((float)degree);
 		}
 	}
-	public void yRotation(float degree){
+	public void yRotation(double degree){
 		for(int i=0;i<pieces.size();i++){
-			pieces.get(i).yRotation(degree);
+			pieces.get(i).yRotation((float)degree);
 		}
 	}
-	public void zRotation(float degree){
+	public void zRotation(double degree){
 		for(int i=0;i<pieces.size();i++){
-			pieces.get(i).zRotation(degree);
+			pieces.get(i).zRotation((float)degree);
 		}
 	}
